fix: change the structure, create authenticationservice and remove the functions from controller
